refactor(map): streamline component structure by replacing inline elements with dedicated components, enhancing readability and maintainability
