+++ src/components/Map/floor-picker-layout.ts
@@ -1,0 +1,8 @@
+export const CELL = 38
+export const GAP = 5
+export const CLOSE = 38
+export const PICKER_TOP = CLOSE + GAP
+export const CONTAINER_TOP = 110 - PICKER_TOP
+
+export const EXPAND_SPRING = { damping: 18, stiffness: 220, mass: 0.8 }
+export const SNAP_SPRING = { damping: 24, stiffness: 280, mass: 0.7 }
