+++ src/components/Map/floor-row.tsx
@@ -1,0 +1,63 @@
+import type React from 'react'
+import { memo } from 'react'
+import { Pressable, Text } from 'react-native'
+import { CELL } from './floor-picker-layout'
+
+interface FloorRowProps {
+	floor: string
+	isCurrent: boolean
+	isLast: boolean
+	interactive: boolean
+	borderColor: string
+	cardColor: string
+	textColor: string
+	contrastColor: string
+	onSelect: (floor: string) => void
+}
+
+export function floorLabel(floor: string): string {
+	return floor === 'EG' ? '0' : floor
+}
+
+export const FloorRow = memo(function FloorRow({
+	floor,
+	isCurrent,
+	isLast,
+	interactive,
+	borderColor,
+	cardColor,
+	textColor,
+	contrastColor,
+	onSelect
+}: FloorRowProps): React.JSX.Element {
+	return (
+		<Pressable
+			testID={`map-floor-${floor}`}
+			onPress={() => {
+				onSelect(floor)
+			}}
+			disabled={!interactive}
+			accessibilityRole="button"
+			accessibilityState={{ selected: isCurrent }}
+			accessibilityLabel={floorLabel(floor)}
+			className="items-center justify-center"
+			style={{
+				height: CELL,
+				width: CELL,
+				backgroundColor: isCurrent && interactive ? 'transparent' : cardColor,
+				borderBottomColor: borderColor,
+				borderBottomWidth: isLast || !interactive ? 0 : 1
+			}}
+		>
+			<Text
+				className="font-medium text-[15px]"
+				style={{
+					color: isCurrent && interactive ? contrastColor : textColor,
+					fontVariant: ['tabular-nums']
+				}}
+			>
+				{floorLabel(floor)}
+			</Text>
+		</Pressable>
+	)
+})
