+++ src/components/Map/sheet-chrome.ts
@@ -6,6 +6,8 @@ export const sheetHostStyle: ViewStyle = {
 	borderTopLeftRadius: SHEET_RADIUS,
 	borderTopRightRadius: SHEET_RADIUS,
 	overflow: 'hidden',
+	zIndex: 40,
+	elevation: 24,
 	...(Platform.OS === 'ios'
 		? {
 				backgroundColor: 'transparent',
